ICH HAB GESAGT DU SOLLST MICH NICHT STRESSEN

+++ NotenOptimierer/backend/run.py
@@ -6,7 +6,6 @@ import json
 def store_cpp_format(parsed_json):
     file_name = "backend/data/"+''.join(random.choices(string.ascii_letters + string.digits, k=8)) + ".data"
     f = open(file_name, "w+")
-    print(parsed_json)
     json_file = open("assets/informatik.json")
     model = json.load(json_file)
     maxDelCredits = model["maxDelCredits"]
@@ -18,8 +17,8 @@ def store_cpp_format(parsed_json):
             if(section_id in parsed_json):
                 section_grades = parsed_json[section_id]
                 subject_ids = section_grades.keys()
-                credits = dict((sub["id"], sub["credits"]) for sub in section["subjects"])
-                f.write(";".join([str(section_grades[sid])+","+str(credits[sid]) for sid in subject_ids]))
+                grade_credits = dict((sub["id"], sub["credits"]) for sub in section["subjects"])
+                f.write(";".join([str(section_grades[sid])+","+str(grade_credits[sid])+","+str(sid) for sid in subject_ids]))
                 f.write("\n")
 
     return file_name
@@ -27,7 +26,11 @@ def store_cpp_format(parsed_json):
 def solve_cpp(parsed_json):
     file_name = store_cpp_format(parsed_json)
     stream = os.popen("./backend/solve " + file_name)
-    output = stream.read()
-    print(output)
+    console_output = stream.read()
+    if console_output:
+        return {"error": console_output}
+    output = []
+    with open(file_name) as f:
+        output = f.read().splitlines()
     os.remove(file_name)
-    return {"ausgabe": output}
+    return {"grade": float(output[0]), "deletions": output[1:]}
